net: cork socket while connecting
Instead of allowing the socket to lazily buffer up a write while
connecting, proactively .cork() the stream and .uncork() when we have
connected. This allows the stream to buffer together all writes while
connecting into a single writev when connected, rather than an initial
write and a follow-up writev.

This only leads us to a small caveat: if the stream is uncorked while
connecting, writes will begin to be sent to the socket unintentionally.
Work around this by preserving a smaller subset of the _pendingData and
_pendingEncoding stuff, which can be used in this case. (This also
happens when .end() is called on a socket which is still connecting.)

This means this will be turned into a single write call:

  var socket = net.connect(...);
  socket.write('hello, wor');
  socket.write('ld!');

Improving performance when this is the case.